refactor: share getSlashCommandMatches from @studio/common
